Fixed a bug in Add Route Activity where distance was not properly updated Fixed a bug in 4-week pie graph where dates where not compared properly Fixed a bug in Route select activity where data did not get updated Fixed a bug in Daily tab where journey emission was not properly calculated

+++ app/src/main/java/com/cmpt276/indigo/carbontracker/RouteAddActivity.java
@@ -97,11 +97,11 @@ public class RouteAddActivity extends AppCompatActivity {
         }
     }
 
-    private int getNumberFromEditTextOrZeroForFail(int id) {
+    private double getNumberFromEditTextOrZeroForFail(int id) {
         EditText data = (EditText) findViewById(id);
         String text = data.getText().toString();
         try {
-            return Integer.parseInt(text);
+            return Double.parseDouble(text);
         } catch (NumberFormatException ex) {
             return 0;
         }
